feat(sso): Propagate `azp` claim
When using Dex with [cross-client trust][cross-client-trust], the `azp`
(Authorized Party) claim is set to the original ID token's `aud`: the
place where the token was issued. This might be used, for example, when
Dex is converting an ID token from GitHub's OIDC.

Make it available in the `Claims` struct, so that RBAC rules can take
advantage of it.

[cross-client-trust]: https://dexidp.io/docs/custom-scopes-claims-clients/#cross-client-trust-and-authorized-party
: