feat(auth): Propagate Dex's `federated_claims` claim
Dex-issued ID tokens have `sub` set to an opaque string. But you can
include the `federated:id` scope to have the upstream provider's `sub`
[returned in `federated_claims`][custom-claims], with the `user_id` key.

This can be useful to have available to RBAC rules as it might contain
information which could be matched on to make decisions. For example,
when using GitHub OIDC the `sub` [by default contains various pieces of
information][github-oidc-subject-claims] about how the job was
triggered, and it can be configured by the repo/org's owner.

[custom-claims]: https://dexidp.io/docs/custom-scopes-claims-clients/#custom-claims
[github-oidc-subject-claims]: https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/about-security-hardening-with-openid-connect#example-subject-claims